Avoid depending on extension from within another extension (#345)
Julia's loading system ought to allow an extension w/ triggers {B,C}
to depend on an extension w/ triggers {A,C}, under the condition that
B depends on A (i.e. the triggers for the second extension are a strict
subset of the triggers for the first extension)

Unfortunately Julia's purely trigger-based mechanism for extensions does
not actually do that. If `C` is the loaded after `{A,B}` then both of
these extensions will trigger "simultaneously" and they will be loaded
in an indeterminate order.

That problem is the same as the "cycle" in
https://github.com/JuliaLang/julia/issues/55557, so-called because pre-
compilation will try to load each of the simultaneously-triggered
extensions "before" the other repeatedly, leading to an ordering cycle.

This works around the problem by removing the inter-extension dependency.